requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
plotly>=5.0.0
kaleido>=0.2.0
